requirements.txt
requests>=2.31.0
pandas>=2.0.0
beautifulsoup4>=4.12.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
tqdm>=4.65.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
nltk>=3.8.0
numpy>=1.24.0
streamlit>=1.28.0
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0
